Extract Item.Explode and Item.Implode
