+++ src/commands/item.js
@@ -2,7 +2,6 @@
 
 const { warn } = require('../common/log')
 const { clipboard, ipcRenderer: ipc } = require('electron')
-const assert = require('assert')
 const { Command } = require('./command')
 const { SaveCommand } = require('./subject')
 const { groupedByTemplate } = require('../export')
@@ -11,7 +10,6 @@ const act = require('../actions')
 const mod = require('../models')
 const { darwin } = require('../common/os')
 const { ITEM } = require('../constants')
-const { keys } = Object
 const { writeFile: write } = require('fs')
 const { win } = require('../window')
 const { get, pluck, remove } = require('../common/util')
@@ -138,99 +136,6 @@ class Split extends Command {
 }
 
 Split.register(ITEM.SPLIT)
-
-
-class Explode extends Command {
-  *exec() {
-    const { db } = this.options
-    const { payload } = this.action
-
-    let item = yield select(state => ({
-      ...state.items[payload.id]
-    }))
-
-    let photos = payload.photos || item.photos.slice(1)
-    let items = {}
-    let moves = {}
-
-    if (payload.items == null) {
-      yield call(db.transaction, async tx => {
-        for (let photo of photos) {
-          let dup = await mod.item.dup(tx, item.id)
-
-          await mod.photo.move(tx, { ids: [photo], item: dup.id })
-          moves[photo] = { item: dup.id }
-          dup.photos.push(photo)
-
-          items[dup.id] = dup
-        }
-      })
-
-      yield put(act.metadata.load(keys(items)))
-
-    } else {
-      items = payload.items
-      let ids = keys(items)
-
-      assert(ids.length === photos.length)
-
-      yield call(db.transaction, async tx => {
-        await mod.item.restore(tx, ids)
-
-        for (let i = 0, ii = photos.length; i < ii; ++i) {
-          let pid = photos[i]
-          let iid = ids[i]
-
-          await mod.photo.move(tx, { ids: [pid], item: iid })
-          moves[pid] = { item: iid }
-        }
-      })
-    }
-
-    yield put(act.photo.bulk.update(moves))
-
-    this.undo = act.item.implode({
-      item, items: keys(items)
-    })
-
-    this.redo = act.item.explode({
-      id: item.id, photos, items
-    })
-
-    this.meta = { inc: photos.length }
-
-    return {
-      ...items,
-      [item.id]: {
-        ...item, photos: remove(item.photos, ...photos)
-      }
-    }
-  }
-}
-
-Explode.register(ITEM.EXPLODE)
-
-
-class Implode extends Command {
-  *exec() {
-    const { db } = this.options
-    const { item, items } = this.action.payload
-    const { id, photos } = item
-
-    yield call(db.transaction, async tx =>
-      mod.item.implode(tx, { id, photos, items }))
-
-    yield put(act.photo.bulk.update([photos, { item: id }]))
-    yield put(act.item.remove(items))
-    yield put(act.item.select({ items: [id] }, { mod: 'replace' }))
-
-    this.meta = { dec: items.length }
-
-    return item
-  }
-}
-
-Implode.register(ITEM.IMPLODE)
 
 
 class Export extends Command {
@@ -432,11 +337,10 @@ ClearTags.register(ITEM.TAG.CLEAR)
 module.exports = {
   ...require('./create'),
   ...require('./delete'),
+  ...require('./explode'),
   ...require('./import'),
 
-  Explode,
   Export,
-  Implode,
   Load,
   Merge,
   Split,
